refactor: improve app architecture and implement document retrieval
