+++ nodemon.json
@@ -1,6 +1,8 @@
 {
-  "watch": ["src"],
+  "watch": [
+    "./src"
+  ],
   "ext": ".ts,.js",
   "ignore": [],
-  "exec": "npx ts-node ./src/app.ts"
+  "exec": "tsx watch src/app.ts"
 }
