+++ src/modules/genai/application/routes/index.ts
