+++ src/modules/genai/core/usecases/search-in-document/types.ts
