+++ src/modules/genai/core/utils/index.ts
